requirements.txt
typer[all]
rich
google-auth-oauthlib
google-api-python-client
